Added a file uploading functionality

+++ src/streamlit_app/manage_projects.py
@@ -1,11 +1,11 @@
 import streamlit as st
 
-
 options = st.session_state.db.list_projects()
 
 st.title("Manage Projects")
 
-with st.popover("Add project"):
+@st.dialog('Add project')
+def add_project():
     name = st.text_input("Project name")
     description = st.text_area("Description")
     if st.button("Add"):
@@ -15,15 +15,23 @@ with st.popover("Add project"):
 
 @st.dialog('Remove project data')
 def remove_project_data():
-    st.write("Are you sure you want to remove all data from this project?")
+    st.write("Remove the project and all the linked data")
     project = st.selectbox("Select project", st.session_state.projects)
-    if st.button('Yes, remove all data', type="primary"):
+    if st.button('Yes, remove the project and all data', type="primary"):
         st.session_state.db.remove_project(project)
-        #st.switch_page("main.py")
         st.session_state.projects = []
         st.session_state.project = ''
         st.rerun()
 
-if len(st.session_state.projects) > 0:
-    if st.button("Remove project data"):
-        remove_project_data()
+col1, col2 = st.columns(2)
+
+
+with col1:
+    if st.button("Add project", use_container_width=True):
+        add_project()
+
+with col2:
+    if len(st.session_state.projects) > 0:
+        if st.button("Remove project", type="primary", use_container_width=True):
+            remove_project_data()
+
